backend/requirements_production.txt
# Production requirements for Chabighar backend - Complete pip freeze
# Deployed at admin.chobighar.com
# Generated on October 2, 2025

aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
altair==5.5.0
amqp==5.3.1
arrow==1.3.0
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.1
binaryornot==0.4.4
blinker==1.9.0
cachetools==6.1.0
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.2
cookiecutter==2.6.0
cryptography==45.0.5
cycler==0.12.1
Django==5.2.6
django-admin-interface==0.30.1
django-appconf==1.1.0
django-browser-reload==1.18.0
django-ckeditor==6.7.3
django-cleanup==8.1.0
django-colorfield==0.14.0
django-compressor==4.4
django-cors-headers==4.9.0
django-debug-toolbar==4.3.0
django-decouple==2.1
django-filter==25.1
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-material==1.12.0
django-material-admin==1.8.6
django-nested-admin==4.1.1
django-nextjs==3.3.0
django-redis==5.4.0
django-rest-framework==0.1.0
django-storages==1.14.2
django-taggit==6.1.0
django-tailwind==4.1.0
django-tinymce==4.1.0
djangorestframework==3.16.1
docx==0.2.4
drf-yasg==1.21.10
factory_boy==3.3.3
Faker==37.3.0
fonttools==4.58.5
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.44
gunicorn==21.2.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
kombu==5.5.4
lxml==6.0.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
multidict==6.6.3
mutagen==1.47.0
mysqlclient==2.2.4
narwhals==1.46.0
nodeenv==1.9.1
numpy==2.3.1
packaging==25.0
pandas==2.3.1
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
pluggy==1.6.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psycopg2-binary==2.9.10
pyarrow==20.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.19.2
PyMuPDF==1.26.3
pyparsing==3.2.3
PyPDF2==3.0.1
pypdfium2==4.30.1
pytesseract==0.3.13
pytest==8.4.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.2.0
python-dotenv==1.1.1
python-magic==0.4.27
python-monkey-business==1.1.0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
rcssmin==1.1.1
redis==6.2.0
referencing==0.36.2
requests==2.32.4
rich==14.0.0
rjsmin==1.2.1
rpds-py==0.26.0
six==1.17.0
smmap==5.0.2
sqlparse==0.5.3
streamlit==1.46.1
tenacity==9.1.2
text-unidecode==1.3
toml==0.10.2
tornado==6.5.1
types-python-dateutil==2.9.0.20250516
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uuid==1.30
vine==5.1.0
watchdog==6.0.0
wcwidth==0.2.13
websockets==15.0.1
whitenoise==6.9.0
yarl==1.20.1